Move to general-tests
