+++ test/e2e/specs/specs-wpcom/wp-calypso-gutenberg-post-editor-tracking-spec.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import config from 'config';
-import assert from 'assert';
 
 /**
  * Internal dependencies
@@ -10,11 +9,10 @@ import assert from 'assert';
 import LoginFlow from '../../lib/flows/login-flow.js';
 
 import WPAdminSidebar from '../../lib/pages/wp-admin/wp-admin-sidebar';
-import GutenbergEditorComponent from '../../lib/gutenberg/gutenberg-editor-component.js';
 
 import * as driverManager from '../../lib/driver-manager.js';
 import * as dataHelper from '../../lib/data-helper.js';
-import { clearEventsStack, getEventsStack } from '../../lib/gutenberg/tracking/utils.js';
+import { clearEventsStack } from '../../lib/gutenberg/tracking/utils.js';
 import { createGeneralTests } from '../../lib/gutenberg/tracking/general-tests.js';
 
 const mochaTimeOut = config.get( 'mochaTimeoutMS' );
@@ -42,19 +40,6 @@ describe( `[${ host }] Calypso Gutenberg Post Editor Tracking: (${ screenSize })
 			it,
 			editorType: 'post',
 			postType: 'post',
-		} );
-
-		it( 'Tracks "wpcom_block_editor_details_open" event', async function () {
-			const editor = await GutenbergEditorComponent.Expect( this.driver, 'wp-admin' );
-
-			await editor.toggleDetails(); // Open details
-			await editor.toggleDetails(); // Close details
-
-			const eventsStack = await getEventsStack( this.driver );
-			const toggleEvents = eventsStack.filter(
-				( [ eventName ] ) => eventName === 'wpcom_block_editor_details_open'
-			);
-			assert.strictEqual( toggleEvents.length, 1 );
 		} );
 
 		afterEach( clearEventsStack );
